requirements.txt
qrcode 
pillow 
